+++ rumqttc/src/eventloop.rs
@@ -331,6 +331,8 @@ pub(crate) async fn socket_connect(
             SocketAddr::V6(_) => TcpSocket::new_v6()?,
         };
 
+        socket.set_nodelay(network_options.tcp_nodelay)?;
+
         if let Some(send_buff_size) = network_options.tcp_send_buffer_size {
             socket.set_send_buffer_size(send_buff_size).unwrap();
         }